Start restructure of code base
