--- src/components/common/index.js
@@ -1,11 +1,0 @@
-import Header from './header';
-import Navbar from './navbar';
-import Searchbar from './searchbar';
-
-
-export {
-    Header,
-    Navbar,
-    Searchbar,
-
-}
